Refactored Folders
Basically the whole point of this is making the 2 teams codes more separated and easier to understand
