+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Examples/Autonomous_Example.java
@@ -1,12 +1,12 @@
-package org.firstinspires.ftc.teamcode.Autonomous;
+package org.firstinspires.ftc.teamcode.Examples;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.firstinspires.ftc.teamcode.Systems.RoadRunner.StandardTrackingWheelLocalizer;
-import org.firstinspires.ftc.teamcode.Systems.Autonomous.*;
-import org.firstinspires.ftc.teamcode.Systems.*;
+import org.firstinspires.ftc.teamcode.Shared.RoadRunner.StandardTrackingWheelLocalizer;
+import org.firstinspires.ftc.teamcode.Shared.Autonomous.*;
+import org.firstinspires.ftc.teamcode.Shared.*;
 
 import java.util.HashMap;
 import java.util.Map;
