+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Robot.java
@@ -1,5 +1,5 @@
 package org.firstinspires.ftc.teamcode;
 
-import org.firstinspires.ftc.teamcode.Robots.*;
+import org.firstinspires.ftc.teamcode.Team201.Robot201;
 
 public interface Robot extends Robot201 {}
