+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Shared/Autonomous/Threaded_Servo.java
@@ -1,7 +1,7 @@
-package org.firstinspires.ftc.teamcode.Systems.Autonomous;
+package org.firstinspires.ftc.teamcode.Shared.Autonomous;
 
 import com.qualcomm.robotcore.hardware.Servo;
-import org.firstinspires.ftc.teamcode.Systems.RobotHardware;
+import org.firstinspires.ftc.teamcode.Shared.RobotHardware;
 
 public class Threaded_Servo extends Thread {
