+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Shared/Logic_Base.java
@@ -1,9 +1,9 @@
-package org.firstinspires.ftc.teamcode.Systems;
+package org.firstinspires.ftc.teamcode.Shared;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.qualcomm.robotcore.hardware.Gamepad;
 
-import org.firstinspires.ftc.teamcode.Systems.RoadRunner.StandardTrackingWheelLocalizer;
+import org.firstinspires.ftc.teamcode.Shared.RoadRunner.StandardTrackingWheelLocalizer;
 import org.firstinspires.ftc.teamcode.Robot;
 
 import java.util.ArrayList;
