+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Shared/RobotHardware.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Systems;
+package org.firstinspires.ftc.teamcode.Shared;
 
 import com.qualcomm.hardware.bosch.BNO055IMU;
 import com.qualcomm.hardware.bosch.JustLoggingAccelerationIntegrator;
