+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Team202/Robot202.java
@@ -1,35 +1,27 @@
-package org.firstinspires.ftc.teamcode.Robots;
+package org.firstinspires.ftc.teamcode.Team202;
 
 import org.firstinspires.ftc.robotcore.external.navigation.AxesOrder;
 import java.util.ArrayList;
 import java.util.Arrays;
 
-public interface Robot201 {
+public interface Robot202 {
     //Game Variables
     double robot_width = 15.0;
     double robot_length = 15.0;
 
-    double[] armPositions = {0, 100, 200, 300, 400}; //Even though motor positions are ints, they have to be doubles
-    ArrayList<String> armPositionNames = new ArrayList<>(Arrays.asList("Intake", "Junction", "Low", "Medium", "High"));
-
-    double[] armPositions2 = {0.0, 0.1, 0.2, 0.3};
-    ArrayList<String> armPosition2Names = new ArrayList<>(Arrays.asList("Intake", "Angled Down", "Straight", "Angled Up"));
-
-    double[] scissorPositions = {0.0, 0.2};
-    ArrayList<String> scissorPositionNames = new ArrayList<>(Arrays.asList("Closed", "Open"));
     //Robot Hardware
 
-    ArrayList<String> dc_motor_names = new ArrayList<>(Arrays.asList("Left", "Right"));
-    double[] max_power = {0.8, 0.8};
-    double[] min_power = {-0.8, -0.8};
+    ArrayList<String> dc_motor_names = new ArrayList<>(Arrays.asList("joint1right", "joint2"));
+    double[] max_power = {0.1256486834658387658, 0.4};
+    double[] min_power = {-0.5, -0.4};
     double[] motor_max_positions = {Double.POSITIVE_INFINITY, Double.POSITIVE_INFINITY};
     double[] motor_min_positions = {Double.NEGATIVE_INFINITY, Double.NEGATIVE_INFINITY};
-    boolean[] invert_dc_motors = {true, false};
-    double[] p_weights = {0.05, 0.05};
+    boolean[] invert_dc_motors = {false, false};
+    double[] p_weights = {0.02, 0.02};
 
-    ArrayList<String> servo_names = new ArrayList<>(Arrays.asList("Virtual", "Scissor"));
+    ArrayList<String> servo_names = new ArrayList<>(Arrays.asList("clawAligner", "claw"));
     double[] servo_max_positions = {1.0, 1.0};
-    double[] servo_min_positions = {0.0, 0.0};
+    double[] servo_min_positions = {0, 0};
 
     ArrayList<String> cr_servo_names = new ArrayList<String>(Arrays.asList());
     boolean[] invert_cr_servos = {};
@@ -43,20 +35,20 @@ public interface Robot201 {
     ArrayList<String> led_names = new ArrayList<>(Arrays.asList());
 
     //Driving
-    double strafe = 0.8;
-    double turning_weight = 1.0;
-    double distance_weight = 1.0;
-    double distance_weight_two = 1.0;
+    double strafe = 0.0;
+    double turning_weight = 0.0;
+    double distance_weight = 0.0;
+    double distance_weight_two = 0.0;
 
-    boolean locked_motion = false; //MAKE THESE TRUE SOON PLS :)
+    boolean locked_motion = true;
     boolean locked_rotation = false;
 
     //PID
     boolean usePID = true;
-    double p_weight = 0; // 0.025
+    double p_weight = 0.025; // 0.025
     double d_weight = 0; // 0.085
 
-    AxesOrder axesOrder = AxesOrder.ZYX;
+    AxesOrder axesOrder = AxesOrder.YXZ;
     boolean invertIMU = false;
 
     //Road Runner
@@ -79,7 +71,7 @@ public interface Robot201 {
             "Vuforia Key";
     String webcam_name = "Webcam 1";
     double camera_zoom = 1.0;
-    String TFOD_MODEL_ASSET = "/sdcard/FIRST/tflitemodels/Sleeve_Detection.tflite"; //Move Sleeve_Detection.tflite to FtcRobotController/assets I think
+    String TFOD_MODEL_ASSET = "PowerPlay.tflite";
     String TFOD_MODEL_FILE = "/sdcard/FIRST/tflitemodels/CustomTeamModel.tflite";
     Float min_confidence = 0.75f;
     int input_size = 300;
@@ -90,7 +82,7 @@ public interface Robot201 {
             "3 Panel"
     };
 
-    ArrayList<String> encoderNames = new ArrayList<>(Arrays.asList("leftEncoder", "rightEncoder", "frontEncoder")); //REV Thorough Bore Encoders
+    ArrayList<String> encoderNames = new ArrayList<>(Arrays.asList("leftEncoder", "rightEncoder", "frontEncoder"));
 
     ArrayList<String> wheel_names = new ArrayList<>(Arrays.asList("rightFront", "rightBack", "leftBack", "leftFront"));
 
@@ -102,5 +94,4 @@ public interface Robot201 {
             "operator left_stick_x", "operator right_stick_x", "operator left_stick_y", "operator right_stick_y",
             "operator left_trigger", "operator right_trigger", "driver left_trigger"
     ));
-
 }
